# [Search] Speed up by 6x with permanent connection pool