--- a/apps/algorithm/search/src/database.py
+++ b/apps/algorithm/search/src/database.py
@@ -1,5 +1,4 @@
 import asyncpg
-from fastapi import HTTPException
 
 from .config import DB_ENDPOINT
 from opentelemetry.instrumentation.asyncpg import AsyncPGInstrumentor
@@ -8,10 +7,12 @@ AsyncPGInstrumentor().instrument()
 
 DB_POOL = None
 
+
 async def initialize_db():
     global DB_POOL
     if not DB_POOL:
         DB_POOL = await asyncpg.create_pool(dsn=DB_ENDPOINT)
+
 
 async def insert_user_search(user_id: int, search_term: str):
     await initialize_db()