@MNThomson 2024-07-16
When combined with #422, gets response time down to ~14ms
![image](https://github.com/user-attachments/assets/2b0d075c-d296-45fa-a914-90c5327e28a8)